soc: octeontx2-rm: add Marvell OcteonTX2 RM driver
commit 25da778fa5fe3825841a9e1ac6649ced2bcb85c0 from
git@git.assembla.com:cavium/WindRiver.linux.git

This patch adds a driver for managing non-NIC hardware offload resources
and exposing them to user through SSO/TIM type of RVU VF devices.
For that reason it acts as a proxy for RVU MBOX messages between SSO/TIM
RVU VFs and RVU AF.

It has previously lived outside of the Marvell internal kernel tree
which is not convenient in development process. Therefore this patch
aims to integrate this driver here with a caveat that upstreaming
process will have to include its re-work.

This driver contains a copy of quota management that is part of RVU AF
driver. That code is not re-used as it would require re-work of the
RVU AF driver which is in process of upstreaming. As part of upstreaming
the resource management interfaces used in this driver will have to
be re-worked too (suggested by upstream maintainers to use devlink) to
accommodate requirements of other users to minimize the functionality of
this driver to an RVU stub.

Change-Id: Ie858a79cf43f30efaaabd8294c0d6464536fe80b
Signed-off-by: Stanislaw Kardach <skardach@marvell.com>
Signed-off-by: Krishnamurthy D V <kvasanthrao@caviumnetworks.com>
Signed-off-by: Xiaotao Yin <xiaotao.yin@windriver.com>