arm64: Add workaround for Marvell erratum 37119
commit a8735dfd13887dac412adf0d8be69904486c0a60 from
git@git.assembla.com:cavium/WindRiver.linux.git

Enable workaround for erratum 37119.  On some OcteonTX2 chips,
while using the SSO workslots in userspace can cause problems
as the cache are not invalidated correctly.  To workaround
this erratum, a local tlbi is placed before the eret.
This TLBI does not need to invalidate anything so a tlbi against
asid 0 is used.

Change-Id: I0313daae25e8ccaaffb84f9b6db5e6859b693dfa
Signed-off-by: Andrew Pinski <apinski@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/17086
Reviewed-by: Stanislaw Kardach <Stanislaw.Kardach@cavium.com>
Reviewed-by: Sunil Kovvuri Goutham <Sunil.Goutham@cavium.com>
Tested-by: Sunil Kovvuri Goutham <Sunil.Goutham@cavium.com>
[Kevin: Use ERRATA_MIDR_RANGE_LIST macro to fix the duplicate entry warning
for capability.]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
Signed-off-by: Xiaotao Yin <xiaotao.yin@windriver.com>