octeontx2-af: Increase mbox response timeout to 2 seconds
commit 3d41081109f29fdf84517b3cd1220852f783b382 from
git@git.assembla.com:cavium/WindRiver.linux.git

It used to be that the intended 1 second mbox response timeout was not very
precise such that the effective timeout was longer than 1 second.  Recently
code was checked in to make the timeout very precise.

It turns out that the old, imprecise, longer-than-1-second timeout was
beneficial.  Now that it's been replaced with a precise 1-second timeout,
we sometimes see these alarming error messages from ethtool:

    Cannot set new settings: Input/output error
    Cannot set private flags: Input/output error

These turn out to be false alarms because the operations (of changing
Ethernet speed or setting private flags) actually take effect.

One second is not enough time for ATF to detect the arrival of a CGX_CMD,
to execute the command, and to send back a response.

Increase the timeout to 2 seconds for ample safety margin.

Change-Id: Ia663fdf55a7a1417fc2d8b9ee751cab0903a7710
Signed-off-by: Felix Manlunas <fmanlunas@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/17678
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <Sunil.Goutham@cavium.com>
(cherry picked from commit b842f3df534a85e70811b4490900a10d17f46f99)
Reviewed-on: https://sj1git1.cavium.com/17824
Tested-by: Sunil Kovvuri Goutham <Sunil.Goutham@cavium.com>
Signed-off-by: Xiaotao Yin <xiaotao.yin@windriver.com>